# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Honcho Inspector Backend
  version: 0.1.0
  description: 'Backend for the Honcho Inspector admin surface.


    Sits between the Angular UI and one or more Honcho instances. Keeps

    the Honcho API key off the browser, externalizes config per-OS

    convention, and serves a multi-user multi-profile admin surface

    (one SQLite DB, no external service to run).


    ## Authentication


    This API does NOT use JWT or OAuth2 Bearer tokens. Every authenticated

    request must carry the `X-Session-Id` header whose value is the opaque

    session id returned by `POST /api/auth/login`. The session id is a

    24-byte random hex string stored server-side; send it verbatim on every

    subsequent request. `POST /api/auth/logout` invalidates it.


    Honcho proxy endpoints additionally require the `X-Honcho-Profile-Id`

    header to pick which encrypted profile (base URL + API key + workspace)

    the request is routed to.


    See `x-workflow-narrative` at the root of this document for the

    recommended end-to-end flow.


    Live Swagger UI: <http://localhost:8080/swagger-ui.html>

    '
  contact:
    name: Honcho Inspector maintainers
    url: https://github.com/cloudbsdorg/honcho-inspector-backend
  license:
    name: BSD-3-Clause
    url: https://opensource.org/licenses/BSD-3-Clause
servers:
- url: http://localhost:8080
  description: Local dev via `mvn spring-boot:run`
- url: https://inspector.example.com
...

...

...

...

    '
...
paths:
...
  /api/admin/users/{id}/sessions:
    get:
      tags:
      - admin
      summary: List a user's auth sessions
      operationId: sessions
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: List of auth sessions (live + recent)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthSession'
...
components:
  schemas:
...
    AuthSession:
      type: object
      description: 'One row of the `auth_sessions` table, returned by `GET

        /api/admin/users/{id}/sessions`.

        '
      properties:
        id:
          type: string
          description: Hex session id (the value sent in `X-Session-Id`)
        userId:
          type: string
          description: Hex user id of the owner
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: 'ISO-8601 instant; null when `SESSION_TTL_MINUTES=0` (never
            expire)'
        ip:
          type: string
          nullable: true
# ----------------------------------------------------------------------------
# Top-level workflow narrative - the recommended end-to-end flow.
# ----------------------------------------------------------------------------
...